packages/scraper/pyproject.toml
[project]
name = "stylai-scraper"
version = "0.1.0"
description = "StylAI - Product scraper for fashion websites"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pydantic>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "typer>=0.15.0",
    "rich>=13.0.0",
    "python-slugify>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
scraper = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
